fix: review white error page
