+++ src/main/resources/static/css/style.css
@@ -110,8 +110,7 @@
     padding: 20px 0;
 }
 
-.navbar-dark .navbar-nav .nav-link:hover,
-.navbar-dark .navbar-nav .nav-link.active {
+.navbar-dark .navbar-nav .nav-link:hover{
     color: var(--primary) !important;
 }
