+++ src/main/resources/templates/auth/login.html
@@ -14,6 +14,8 @@
 <link rel="stylesheet"
 	href="https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@400;500;700&display=swap">
 
+<th:block th:insert="~{/header_unlogin.html}" />
+
 <style>
 :root {
 	--primary: #FEA116;
@@ -70,6 +72,10 @@ body {
 .btn-primary:hover {
 	background-color: var(--primary-hover);
 	border-color: var(--primary-hover);
+}
+
+h3.h5, i.bi{
+	color: var(--primary)
 }
 </style>
 </head>
